refactor(package.json): updating files
