--- asm.js
@@ -1,6 +1,0 @@
-'use strict';
-
-const loadSVM = require('./src/loadSVM');
-const libsvm = require('./out/asm/libsvm');
-
-module.exports = loadSVM(libsvm);
