+++ benchmark/browser.js
@@ -3,7 +3,7 @@ export default async function exec(benchmark, mode, time) {
     if(mode === 'asm') {
         SVM = await import('../asm');
     } else if(mode === 'wasm') {
-        SVM = await import('../wasm');
+        SVM = await import('..');
     } else {
         throw new Error('Invalid mode');
     }
